chore(core): clean up unit tests and finalize C++ core structure
- Removed redundant test_integration.cpp and generate_fixtures.py.
- Simplified docker-compose.yml and Makefile for pure library unit testing.
- Fixed public visibility of unpackISRC in MetadataDatabase.h.
- Verified stable pass rate for Art, Meta, and Searcher modules.

+++ libmusicprint/include/Searcher.h
@@ -8,7 +8,7 @@
 namespace musicprint {
 
 struct SearchResult {
-    uint32_t song_id;
+    uint64_t song_id; // Packed ISRC
     float distance;
 };
